chore: restructure tests to map to source files
 - the test names didn't correspond to files and it made it difficult to know where a particular test might be.
