+++ tests/test_client.py
@@ -10,13 +10,6 @@ from gcspath import PureGCSPath
 from gcspath.file import BucketClientFS, ClientBlobFS, ClientBucketFS
 
 
-@pytest.fixture()
-def temp_folder():
-    tmp_dir = tempfile.mkdtemp()
-    yield Path(tmp_dir)
-    shutil.rmtree(tmp_dir)
-
-
 def test_client_create_bucket(temp_folder: Path):
     bucket_target = temp_folder / "foo"
     assert bucket_target.exists() is False
